requirements.txt
feedparser==6.0.11
groq==0.4.2
httpx==0.27.2
streamlit==1.30.0
python-dotenv==1.0.1
streamlit-lottie==0.0.4